--- a/collector/series.go
+++ b/collector/series.go
@@ -9,30 +9,30 @@ type seriesCreator struct{}
 
 func (s *seriesCreator) newSeries(name string, scrapeTarget ScrapeTarget, m *io_prometheus_client.Metric) prompb.TimeSeries {
 	ts := prompb.TimeSeries{
-		Labels: make([]prompb.Label, 0, len(m.Label)+3),
+		Labels: make([]*prompb.Label, 0, len(m.Label)+3),
 	}
 
-	ts.Labels = append(ts.Labels, prompb.Label{
+	ts.Labels = append(ts.Labels, &prompb.Label{
 		Name:  []byte("__name__"),
 		Value: []byte(name),
 	})
 
 	if scrapeTarget.Namespace != "" {
-		ts.Labels = append(ts.Labels, prompb.Label{
+		ts.Labels = append(ts.Labels, &prompb.Label{
 			Name:  []byte("adxmon_namespace"),
 			Value: []byte(scrapeTarget.Namespace),
 		})
 	}
 
 	if scrapeTarget.Pod != "" {
-		ts.Labels = append(ts.Labels, prompb.Label{
+		ts.Labels = append(ts.Labels, &prompb.Label{
 			Name:  []byte("adxmon_pod"),
 			Value: []byte(scrapeTarget.Pod),
 		})
 	}
 
 	if scrapeTarget.Container != "" {
-		ts.Labels = append(ts.Labels, prompb.Label{
+		ts.Labels = append(ts.Labels, &prompb.Label{
 			Name:  []byte("adxmon_container"),
 			Value: []byte(scrapeTarget.Container),
 		})
@@ -43,7 +43,7 @@ func (s *seriesCreator) newSeries(name string, scrapeTarget ScrapeTarget, m *io_
 			continue
 		}
 
-		ts.Labels = append(ts.Labels, prompb.Label{
+		ts.Labels = append(ts.Labels, &prompb.Label{
 			Name:  []byte(l.GetName()),
 			Value: []byte(l.GetValue()),
 		})