--- a/ingestor/service_test.go
+++ b/ingestor/service_test.go
@@ -93,7 +93,7 @@ func (f fakeStore) Close() error {
 	panic("implement me")
 }
 
-func (f fakeStore) WriteTimeSeries(ctx context.Context, ts []prompb.TimeSeries) error {
+func (f fakeStore) WriteTimeSeries(ctx context.Context, ts []*prompb.TimeSeries) error {
 	panic("implement me")
 }