--- a/pkg/prompb/util.go
+++ b/pkg/prompb/util.go
@@ -1,6 +1,6 @@
 package prompb
 
-func MetricName(ts TimeSeries) []byte {
+func MetricName(ts *TimeSeries) []byte {
 	for _, l := range ts.Labels {
 		if string(l.Name) == "__name__" {
 			return l.Value