This is a followup to the prior streaming collection PR where we reduce the CPU and allocations incurred during scraping.

benchmark                           old ns/op     new ns/op     delta
BenchmarkIterator_TimeSeries-10     365           119           -67.39%

benchmark                           old allocs     new allocs     delta
BenchmarkIterator_TimeSeries-10     9              0              -100.00%

benchmark                           old bytes     new bytes     delta
BenchmarkIterator_TimeSeries-10     480           0             -100.00%